requirements.txt
Flask
whisper-openai
flask-cors